cp-38: * split stack and bottom tabs navigation
