+++ mobile/src/libs/enums/navigation/root-screen-name.enum.ts
@@ -1,8 +1,8 @@
 const RootScreenName = {
-  SIGN_UP: 'Sign Up',
-  SIGN_IN: 'Sign In',
-  MAIN: 'Main Screens',
   HOME: 'Home',
+  MAIN: 'Main Screens',
+  SIGN_IN: 'Sign In',
+  SIGN_UP: 'Sign Up',
 } as const;
 
 export { RootScreenName };
