+++ mobile/src/libs/types/navigation/root-navigation-parameter-list.type.ts
@@ -1,9 +1,9 @@
 import { type RootScreenName } from '#libs/enums/enums';
 
 type RootNavigationParameterList = {
+  [RootScreenName.MAIN]: undefined;
   [RootScreenName.SIGN_IN]: undefined;
   [RootScreenName.SIGN_UP]: undefined;
-  [RootScreenName.MAIN]: undefined;
 };
 
 export { type RootNavigationParameterList };
