+++ mobile/src/navigations/root/main/styles.ts
